Created class s3Agent for s3Client
